+++ src/main/java/com/pkware/generex/GenerexIterator.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mifmif.common.regex;
+package com.pkware.generex;
 
 import dk.brics.automaton.State;
 import dk.brics.automaton.Transition;
