+++ src/main/java/com/pkware/generex/Node.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.mifmif.common.regex;
+package com.pkware.generex;
 
 import java.util.ArrayList;
 import java.util.List;
