+++ src/test/java/com/pkware/generex/GenerexIteratorTest.java
@@ -1,4 +1,4 @@
-package com.mifmif.common.regex;
+package com.pkware.generex;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
